fix type

+++ src/motor.cpp
@@ -11,10 +11,10 @@ void MOTORS::stop(int MOTOR_ID){
 void MOTORS::stop_all(){
 
 }
-void MOTORS::set_power_to(int MOTOR_ID , int Power){
+void MOTORS::set_power_to(int MOTOR_ID , int16_t Power){
     power[MOTOR_ID] = Power;
 }
 
-void MOTORS::set_power_all(int Power[]){
+void MOTORS::set_power_all(int16_t Power[]){
 
 }
